Add NNeighbors test

+++ spectraclass/test/nearest_neighbors_test-3.py
@@ -8,7 +8,7 @@ from cudf.io.csv import read_csv
 print_rows = 25
 n_neighbors = 4
 
-device_data: cudf.DataFrame = read_csv( "spectraclass_data.csv", dtype=np.float32 ).pivot(index='samples', columns='model')
+device_data: cudf.DataFrame = read_csv( "spectraclass_data.csv", dtype=np.float64 ).pivot(index='samples', columns='model')
 host_data: pd.DataFrame = device_data.to_pandas()
 print( f"\nINPUT DATA:\n{device_data.head(10)}")
 print( f"Host dataframe: ndim = {host_data.ndim}, size = {host_data.size}, col shape = {host_data.columns.shape}, shape = {host_data.index.shape}, np-shape = {host_data.values.shape}" )
